Resolved circural dependency
