+++ .github/workflows/static-root-checks.yml
@@ -29,3 +29,7 @@ jobs:
         run: yarn tsc --noEmit
       - name: Lint
         run: yarn lint:js
+      - name: Finding circular dependencies
+        run: yarn circural_dependency_check
+      - name: Jest test:unit
+        run: yarn test:unit
